Move ExoFOP retrieval to TESS TOI data interface
